Fixing up functions

+++ doubly_linked_list/doubly_linked_list.py
@@ -104,37 +104,43 @@ class DoublyLinkedList:
     """Removes the input node from its current spot in the 
     List and inserts it as the new tail node of the List."""
     def move_to_end(self, node):
-        prev_node = node.prev
-        next_node = node.next
+        if node is self.tail:
+            return
 
-        prev_node.next = next_node
-        next_node.prev = prev_node
-
+        value = node.value
+        self.delete()
         self.add_to_tail(node)
 
     """Removes a node from the list and handles cases where
     the node was the head or the tail"""
     def delete(self, node):
-        if node == self.head:
+        self.length -= 1
+
+        if not self.head and not self.tail:
+            return
+        elif self.head == self.tail:
+            self.head = None
+            self.tail = None
+        elif node == self.head:
             self.head = self.head.next
+            node.delete()
         elif node == self.tail:
             self.tail = self.tail.prev
+            node.delete()
         else:
-            prev_node = node.prev
-            next_node = node.next
-
-            prev_node.next = next_node
-            next_node.prev = prev_node
-
-        self.delete()
-        self.length -= 1
+            node.delete()
         
     """Returns the highest value currently in the list"""
     def get_max(self):
+        if not self.head:
+            return None
+
         max_val = 0
         current = self.head
 
         while current:
             if current.value > max_val:
                 max_val = current
-            current = self.head.next
+            current = current.next
+
+        return max_val
